pyproject.toml
[tool.poetry]
name = "hw_31"
version = "0.1.0"
description = ""
authors = ["Maxim <99118327+Maximz2@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
